defined entity and repository
