+++ src/modules/pokemon/pokemon.repository.ts
@@ -1,44 +1,18 @@
-import { Entity, PrimaryColumn, Column } from 'typeorm';
+import { Injectable } from '@nestjs/common';
+import { InjectRepository } from '@nestjs/typeorm';
+import { Repository } from 'typeorm';
+import { Pokemon } from './pokemon.entity';
 
-@Entity({ name: 'pokemon' })
-// TODO: i might want to define english name of a Pokemon as Unique
-export class Pokemon {
-  @PrimaryColumn({ type: 'int' })
-  id: number;
-
-  @Column({ type: 'varchar', length: 50 })
-  nameEnglish: string;
-
-  @Column({ type: 'simple-array' })
-  type: string[];
-
-  @Column({ type: 'jsonb' })
-  base: {
-    HP: number;
-    Attack: number;
-    Defense: number;
-    Speed: number;
-  };
-  @Column({ type: 'varchar', length: 255 })
-  description: string;
-
-  @Column({ type: 'jsonb' })
-  image: {
-    sprite: string;
-    thumbnail: string;
-    hires: string;
-  };
-
-  @Column({ type: 'simple-array' })
-  weight: number;
-
-  @Column({ type: 'varchar', length: 50 })
-  species: number;
-
-  @Column({ type: 'jsonb' })
-  profile: {
-    height: string;
-    weight: string;
-    ability: string[][];
-  };
+@Injectable()
+export class PokemonRepository {
+  constructor(
+    @InjectRepository(Pokemon)
+    private readonly pokemonRepository: Repository<Pokemon>,
+  ) {}
+  async findAll(): Promise<Pokemon[]> {
+    return this.pokemonRepository.find();
+  }
+  async findOneByID(id: number) {
+    return this.pokemonRepository.findOneBy({ id });
+  }
 }
